requirements.txt
psycopg2-binary
fastapi
uvicorn
chromadb
beautifulsoup4
googletrans
langdetect
pytest
ruff
python-dotenv
sentence-transformers
langchain
langchain-core
langchain-community
langchain-openai
langgraph
pydantic>=2.0
langgraph-checkpoint-mongodb